added mouse controll and refactoring
