+++ package.json
@@ -1,10 +1,11 @@
 {
   "name": "cli-active-table",
-  "version": "1.1.3",
+  "version": "1.1.4",
   "description": "cli tool that helps you render data in command line interface",
   "main": "dist/active-table.js",
   "scripts": {
     "build": "rm -rf dist & tsc",
+    "test": "tsx tests/index.ts",
     "start": "tsx examples/index.ts",
     "examples:multiple": "tsx examples/multiple.ts",
     "examples:issues": "tsx examples/issues.ts",
